requirements.txt
openai
requests
python-dotenv
streamlit
pandas
sentence-transformers
scikit-learn
firebase-admin
